Try unlocking keychain before `make binary-dist`
I believe the keychain problems we're having now is due to us trying to use keychain objects in `make binary-dist`, and it worked so far because the keychains were remaining unlocked from previous CI runs.  😳  After the buildbots were restarted, everything stonewalled on the `make binary-dist` step, and the keychains were never unlocked.